+++ utils/exp_md5.h
@@ -36,7 +36,7 @@ std::string FileMd5Ex(LPCSTR pszFileName, int64_t maxLen) {
 		{
 			if (maxLen > 0 && pos >= maxLen)
 				break;
-			int nReaded = fread(buf, 1, 1024, f);
+			int nReaded = (int)fread(buf, 1, 1024, f);
 			pos += nReaded;
 			if (nReaded > 0) {
 				MD5Update(&context, buf, nReaded);
